Move handler logic outside of controller
